refactor(scan): extract state scan into its own module
scan/mod.rs was 689 lines, in the Code Quality Way's Review tier (>500).
Pure structural split — `pub fn state` and its private helpers
(`evaluate_context_threshold`, `detect_window_chars`,
`transcript_size_since_summary`, `evaluate_file_exists`,
`strip_frontmatter`, `capture_show_core`) move verbatim into
`scan/state.rs`. The state path was the natural seam: 200+ lines of
state-trigger evaluation that share no helpers with the prompt /
command / file / task scan paths apart from `emit_hook_context`.

`emit_hook_context` stays in mod.rs (still used by `prompt`) and is
re-scoped from private to `pub(super)` so state.rs can route through
it. `pub use state::state` keeps the call site `cmd::scan::state` —
no churn for main.rs or anywhere else.

After split:
- mod.rs: 485 lines (below the Review threshold)
- state.rs: 218 lines (new)

Behavior unchanged — 11/11 sim and 70/70 unit tests pass.

Refs: #81 (item 3)
